Bug fix from original DOS 1:
DOS 1 supports AUTO.SYS, which should be the same as AUTORUN.SYS in DOS 2, except that the binary load header is 84 09 instead of FF FF.  But it only processed a run address when loading files from the menu.  So AUTO.SYS would be loaded, but not executed.  The fix was to put the execution of a run address into the common load file code.